# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: B2B AI Marketing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List users
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: role
          in: query
          schema:
            type: ["string", "null"]
            enum: [super_admin, client_admin, dept_manager, end_user, null]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [active, inactive, at_risk, pending, null]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create / invite a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (role, department, status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete / deactivate a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/bulk-invite:
    post:
      operationId: bulkInviteUsers
      tags: [users]
      summary: Bulk invite users by email list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkInviteInput"
      responses:
        "200":
          description: Bulk invite result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkInviteResult"

  # ── Licenses ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, status, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, client_admin, dept_manager, end_user]
        status:
          type: string
          enum: [active, inactive, at_risk, pending]
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        seatId:
          type: ["integer", "null"]
        lastActiveAt:
          type: ["string", "null"]
        inactiveDays:
          type: ["integer", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [email, name, role]
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, client_admin, dept_manager, end_user]
        clientId:
          type: ["integer", "null"]
        departmentId:
          type: ["integer", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
          enum: [super_admin, client_admin, dept_manager, end_user, null]
        status:
          type: ["string", "null"]
          enum: [active, inactive, at_risk, pending, null]
        departmentId:
          type: ["integer", "null"]
    BulkInviteInput:
      type: object
      required: [emails, role, clientId]
      properties:
        emails:
          type: array
          items:
            type: string
        role:
          type: string
          enum: [client_admin, dept_manager, end_user]
        clientId:
          type: integer
        departmentId:
          type: ["integer", "null"]
    BulkInviteResult:
      type: object
      required: [invited, failed, total]
      properties:
        invited:
          type: integer
        failed:
          type: integer
        total:
          type: integer
        errors:
          type: array
          items:
            type: string

    # ── License Schemas ────────────────────────────────────────────────────────
...