# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * B2B AI Marketing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';
import type { UserUpdateStatus } from './userUpdateStatus';

export interface UserUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  role?: UserUpdateRole;
  /** @nullable */
  status?: UserUpdateStatus;
  /** @nullable */
  departmentId?: number | null;
}